feat(tools): move ToolProvider out of experimental namespace
Move the ToolProvider abstract base class from strands.experimental.tools
to strands.tools as a stable public API with backward compatibility.

Changes:
- Move ToolProvider class to src/strands/tools/tool_provider.py
- Export ToolProvider from strands.tools namespace
- Add deprecation warning for old import path (strands.experimental.tools)
- Update internal imports to use new path
- Remove experimental warning from MCPClient docstring
- Update AGENTS.md directory structure documentation

The old import path still works but emits a DeprecationWarning
advising users to import from strands.tools instead.

Resolves: #1563
