+++ src/strands/tools/__init__.py
@@ -5,6 +5,7 @@ This module provides the core functionality for creating, managing, and executin
 
 from .decorator import tool
 from .structured_output import convert_pydantic_to_tool_spec
+from .tool_provider import ToolProvider
 from .tools import InvalidToolUseNameException, PythonAgentTool, normalize_schema, normalize_tool_spec
 
 __all__ = [
@@ -14,4 +15,5 @@ __all__ = [
     "normalize_schema",
     "normalize_tool_spec",
     "convert_pydantic_to_tool_spec",
+    "ToolProvider",
 ]
