+++ src/strands/tools/tool_provider.py
@@ -5,7 +5,7 @@ from collections.abc import Sequence
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
-    from ...types.tools import AgentTool
+    from ..types.tools import AgentTool
 
 
 class ToolProvider(ABC):
